local-be/requirements.txt
fastapi
elasticsearch
sentence-transformers
numpy
python-dotenv
uvicorn